making AbstractGoogleRobotCredentialsDescriptor inner class of GoogleRobotCredentials
